requirements.txt
bcrypt==4.0.1
cryptography
fastapi
psycopg2-binary
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib
types-passlib
pytest
httpx
pytest-asyncio
aiofiles
jinja2
alembic
python-jose
types-python-jose